+++ instrumentation/CHANGELOG.md
@@ -3,6 +3,8 @@ Change Log
 
 ## Unreleased
 
+## 1.5.0 (2024-07-25)
+
 - Fix inheritance hierarchy of `ComposeExtension` to avoid false-positive warning regarding `@RegisterExtension` (#318)
 - Improve parallel test execution for Android instrumentation tests
 - Fix invalid naming of dynamic tests when executing only a singular test method from the IDE (#317, #339)